feat: a few notebooks analyzing the datasets have been posted
The datasets themselves have not been uploaded due to concerns
on repository storage space. They can be zipped and uploaded.
It is also to be noted that the sepehr2.0 data dictionary is
present in the data folder and can be used to understand the
attributes that we may possess and that can be used for visualizing
trends. (provided that they are not "NA" fields).

The observations.md contains some thoughts that I went through
while researching datasets. The latest notebook that I've used
to analyze the datasets that may be relevant for our project is
the `analyzing_sepher.ipynb` notebook.